+++ R/dependencies.R
@@ -3,7 +3,12 @@
 #' @description Intended for debugging and checking your project.
 #'   The dependency structure of the components of your analysis
 #'   decides which targets are built and when.
-#' @details If the argument is a `knitr` report
+#' @details
+#'   The `globals` slot of the output list contains candidate globals only.
+#'   Each global will be treated as an actual dependency if and only if
+#'   it is either a target or an item in the `envir` argument to [make()].
+#'
+#'   If the argument is a `knitr` report
 #'   (for example, `file_store("report.Rmd")` or `"\"report.Rmd\""`)
 #'   the the dependencies of the expected compiled
 #'   output will be given. For example, `deps_code(file_store("report.Rmd"))`
@@ -30,6 +35,9 @@
 #' @param x a language object (code), character string (code as text),
 #'   or imported function to analyze for dependencies.
 #' @return Names of dependencies listed by type (object, input file, etc).
+#'   The `globals` slot of the output list contains candidate globals only.
+#'   Each global will be treated as an actual dependency if and only if
+#'   it is either a target or an item in the `envir` argument to [make()].
 #' @examples
 #' # Your workflow likely depends on functions in your workspace.
 #' f <- function(x, y){
@@ -79,6 +87,7 @@ deps_code <- function(x){
 #' @description Intended for debugging and checking your project.
 #'   The dependency structure of the components of your analysis
 #'   decides which targets are built and when.
+#' @seealso deps_code
 #' @export
 #' @param target a symbol denoting a target name, or if `character_only`
 #'   is TRUE, a character scalar denoting a target name.
@@ -255,8 +264,8 @@ nonfile_target_dependencies <- function(targets, config, jobs = 1){
   intersect(out, config$plan$target)
 }
 
-import_dependencies <- function(expr, exclude = character(0)){
-  deps <- code_dependencies(expr, exclude = exclude)
+import_dependencies <- function(expr, exclude = character(0), globals = NULL){
+  deps <- code_dependencies(expr, exclude = exclude, globals = globals)
   # Imported functions can't have file_out() deps # nolint
   # or target dependencies from knitr code chunks.
   # However, file_in()s are totally fine. # nolint
@@ -264,12 +273,20 @@ import_dependencies <- function(expr, exclude = character(0)){
   deps
 }
 
-command_dependencies <- function(command, exclude = character(0)){
+command_dependencies <- function(
+  command,
+  exclude = character(0),
+  globals = NULL
+){
   if (!length(command)){
     return()
   }
   command <- as.character(command)
-  deps <- code_dependencies(parse(text = command), exclude = exclude)
+  deps <- code_dependencies(
+    parse(text = command),
+    exclude = exclude,
+    globals = globals
+  )
   deps$strings <- NULL
 
   # TODO: this block can go away when `drake`
@@ -364,7 +381,7 @@ unwrap_function <- function(funct){
   funct
 }
 
-code_dependencies <- function(expr, exclude = character(0)){
+code_dependencies <- function(expr, exclude = character(0), globals = NULL){
   if (
     !is.function(expr) &&
     !is.expression(expr) &&
@@ -386,7 +403,8 @@ code_dependencies <- function(expr, exclude = character(0)){
       }
       walk(body(expr))
     } else if (is.name(expr)) {
-      new_globals <- setdiff(x = wide_deparse(expr), y = drake_fn_patterns)
+      new_globals <- setdiff(x = wide_deparse(expr), y = drake_fn_patterns) %>%
+        Filter(f = is_parsable)
       results$globals <<- c(results$globals, new_globals)
     } else if (is.character(expr)) {
       results$strings <<- c(results$strings, expr)
@@ -416,14 +434,16 @@ code_dependencies <- function(expr, exclude = character(0)){
   }
   walk(expr)
   results$globals <- intersect(results$globals, find_globals(expr))
-  if (length(exclude) > 0){
-    results <- lapply(
-      X = results,
-      FUN = function(x){
-        setdiff(x, exclude)
-      }
-    )
+  if (!is.null(globals)){
+    results$globals <- intersect(results$globals, globals)
   }
+  exclude <- base::union(exclude, ".")
+  results <- lapply(
+    X = results,
+    FUN = function(x){
+      setdiff(x, exclude)
+    }
+  )
   results[purrr::map_int(results, length) > 0]
 }