+++ R/envir.R
@@ -105,7 +105,7 @@ flexible_get <- function(target, envir) {
   lang <- parsed[[1]]
   is_namespaced <- length(lang) > 1
   if (!is_namespaced){
-    return(get(x = target, envir = envir))
+    return(get(x = target, envir = envir, inherits = FALSE))
   }
   stopifnot(deparse(lang[[1]]) %in% c("::", ":::"))
   pkg <- deparse(lang[[2]])